+++ Gemfile
@@ -71,6 +71,7 @@ gem 'after_commit_everywhere', '~> 1.2', '>= 1.2.1'
 gem 'sidekiq', '~> 6.5.7'
 gem 'sinatra', '~> 3.0.2', require: false
 gem 'slim', '~> 4.1.0'
+gem 'sidekiq-cron'
 
 # for authorization
 gem 'cancancan', '~> 3.4.0'
