+++ app/services/facilities_management/admin/management_report_csv_generator.rb
@@ -1,0 +1,30 @@
+module FacilitiesManagement::Admin
+  class ManagementReportCsvGenerator
+    def initialize(id)
+      @management_report = service_name::ManagementReport.find(id)
+      @start_date = @management_report.start_date
+      @end_date = @management_report.end_date
+    end
+
+    def generate
+      report_csv = StringIO.new service_name::ProcurementCsvExport.call(@start_date, @end_date)
+      @management_report.management_report_csv.attach(io: report_csv, filename: "procurements_data_#{created_at}_#{date_to_string(@start_date)}-#{date_to_string(@end_date)}.csv", content_type: 'text/csv')
+      @management_report.complete
+      @management_report.save
+    end
+
+    private
+
+    def date_to_string(date)
+      date&.in_time_zone('London')&.strftime '%Y%m%d'
+    end
+
+    def created_at
+      @management_report.created_at&.in_time_zone('London')&.strftime '%Y%m%d-%H%M'
+    end
+
+    def service_name
+      self.class.module_parent
+    end
+  end
+end
